orgernized my zsh stuff

+++ zsh/aliases.zsh
@@ -1,7 +1,4 @@
 # Basic Alias's for unix and windows
-alias ls='ls -F --color=auto'
-alias la='ls -aF --color=auto'
-alias ll='ls -laF --color=auto'
 alias q='exit'
 alias rma='sudo rm -rv'
 alias ..='cd ../'
@@ -11,21 +8,12 @@ alias .....='cd ../../../../'
 alias ex='chmod +x'
 alias so='source ~/dotfiles/zsh/.zshrc && espanso restart'
 alias mk='mkdir'
-alias file='nautilus'
 alias to='touch'
 alias c='cd $(dirname "$(fzf -e)")'
 alias dl='youtube-dl -x -o "%(title)s.%(ext)s"'
-alias copy='xclip -selection "primary" <'
-alias t='tree -I node_modules'
 alias screenkey='screenkey -p fixed -g 50%x5%+50%+95%'
-alias tags='ctags -R --sort=1 --exclude=node_modules'
 alias v='nvim'
 alias f='rg --files | fzf | xargs -I "{}" nvim {}'
-alias emacs='emacs -l ~/dotfiles/emacs/init.el'
-alias edit='code .&& exit'
-alias -s js=code
-alias vfn='v ~/.config/nvim/functions.vim'
-alias vid='mpv'
 
 
 # Config Files
@@ -33,10 +21,12 @@ alias vrc='v ~/.config/nvim/init.vim'
 alias vma='v ~/.config/nvim/mappings.vim'
 alias vpl='v ~/.config/nvim/plugins.vim'
 alias vpc='v ~/.config/nvim/plugin-configs.vim'
+alias vfn='v ~/.config/nvim/functions.vim'
+alias vsn='v ~/.config/nvim/lua/nvim-snippets.lua'
+alias vst='v ~/.config/nvim/steno.vim'
 alias irc='v ~/.config/i3/config'
 alias pol='v ~/.config/polybar/config'
 alias zsh='v ~/dotfiles/zsh/.zshrc'
-alias vst='v ~/.config/nvim/steno.vim'
 alias val='v ~/.config/plover/vim.json'
 alias kit='v ~/.config/kitty/kitty.conf'
 alias sxh='v ~/.config/sxhkd/sxhkdrc'
@@ -44,7 +34,6 @@ alias bsp='v ~/.config/bspwm/bspwmrc'
 alias mux='v ~/.config/tmux/.tmux.conf'
 alias vif='v ~/.config/vifm/vifmrc'
 alias lrc='v ~/.config/nvim/lua/init.lua'
-alias vsn='v ~/.config/nvim/lua/nvim-snippets.lua'
 alias ala='v ~/.config/alacritty/alacritty.yml'
 
 # Diretorys
@@ -66,25 +55,40 @@ alias esp='cd ~/dotfiles/espanso'
 alias nots='v ~/dotfiles/notes/topics.md'
 alias nodejs='cd ~/code/nodejs'
 
-
 # Linux
-uefi() {
-  sudo systemctl reboot --firmware-setup
-}
-alias boot='reboot'
-alias back='kill -9 -1'
-alias aur='makepkg -si'
-alias sleep='systemctl suspend'
-# alias pacman='sudo pacman'
-alias power='poweroff'
-alias i3='cd i3'
-alias update='yay -Syu; sudo pacman -Sc'
-alias pacman-keys='gpg --keyserver pool.sks-keyservers.net --recv-keys'
-alias usb='cd /run/media/malachi'
+if [[ "$OSTYPE" == "linux-gnu"* ]]
+then
+  alias ls='ls -F --color=auto'
+  alias la='ls -aF --color=auto'
+  alias ll='ls -laF --color=auto'
+  alias power='poweroff'
+  alias update='yay -Syu; sudo pacman -Sc'
+  # alias usb='cd /run/media/malachi'
+  alias boot='reboot'
+  alias back='kill -9 -1'
+  alias aur='makepkg -si'
+  alias sleep='systemctl suspend'
+  alias pacman='sudo pacman'
+  alias pacman-keys='gpg --keyserver pool.sks-keyservers.net --recv-keys'
+  alias vid='mpv'
+  alias copy='xclip -selection "primary" <'
+  uefi() {
+    sudo systemctl reboot --firmware-setup
+    alias file='nautilus'
+  }
+fi
+
+# Mac OS
+if [[ "$OSTYPE = darwin" ]]
+  alias ls='ls -FG'
+  alias la='ls -aFG'
+  alias ll='ls -laFG'
+then
+fi
 
 # Windows
-if [ -d "/mnt/c/Users" ]
-  then
+if [[ "$OSTYPE = win32" ]]
+then
   alias win='cd /mnt/c/Users/malac'
 fi
